backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "DevSync AI Backend — FastAPI + Claude orchestration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "openai-agents>=0.0.17",
    "litellm[google]>=1.55",
    "anthropic>=0.50",
    "httpx>=0.27",
    "asyncpg>=0.30",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]
